+++ packages/connect/CHANGELOG.md
@@ -21,3 +21,4 @@
 
 -   `getOwnershipId` method
 -   `getOwnershipProof` method
+-   `authorizeCoinJoin` method